+++ tests/rest/test_connection.py
@@ -4010,3 +4010,10 @@ class TestMultiResultHandling:
                 "result": True,
             },
         }
+
+
+def test_webeditor(requests_mock):
+    requests_mock.get(API_URL, json=build_capabilities())
+    con = Connection(API_URL)
+    assert con.webeditor() == "https://editor.openeo.org/?server=https%3A%2F%2Foeo.test%2F"
+    assert con.webeditor(anonymous=True) == "https://editor.openeo.org/?server=https%3A%2F%2Foeo.test%2F&discover=1"